created a constant page to hold the object arrays
