+++ src/NewPAge/CrewPage.js
@@ -1,39 +1,6 @@
 import React, { useState, useEffect, useReducer } from "react";
+import { crews } from "../components/Constant";
 
-const crews = [
-  {
-    id: "01",
-    position: "COMMANDER",
-    name: "DOUGLAS HURLEY",
-    about:
-      "Douglas Gerald Hurley is an American engineer, former Marine Corps pilot and former NASA astronaut. He launched into space for the third time as commander of Crew Dragon Demo-2.",
-    photo: "/assets/crew/image-douglas-hurley.png",
-  },
-  {
-    id: "02",
-    position: "MISSION SPECIALIST",
-    name: "MARK SHUTTLEWORTH",
-    about:
-      "Mark Richard Shuttleworth is the founder and CEO of Canonical, the company behind the Linux-based Ubuntu operating system. Shuttleworth became the first South African to travel to space as a space tourist.",
-    photo: "/assets/crew/image-mark-shuttleworth.webp",
-  },
-  {
-    id: "03",
-    position: "PILOT",
-    name: "VICTOR GLOVER",
-    about:
-      "Pilot on the first operational flight of the SpaceX Crew Dragon to the International Space Station. Glover is a commander in the U.S. Navy where he pilots an F/A-18.He was a crew member of Expedition 64, and served as a station systems flight engineer. ",
-    photo: "/assets/crew/image-victor-glover.webp",
-  },
-  {
-    id: "04",
-    position: "FlIGHT ENGINEER",
-    name: "ANOUSHEH ANSARI",
-    about:
-      "Anousheh Ansari is an Iranian American engineer and co-founder of Prodea Systems. Ansari was the fourth self-funded space tourist, the first self-funded woman to fly to the ISS, and the first Iranian in space. ",
-    photo: "/assets/crew/image-anousheh-ansari.webp",
-  },
-];
 export default function CrewPage() {
   const [selectedIndex, setSelectedIndex] = useState(0);
   const ref = [];
