+++ src/NewPAge/Destination.js
@@ -1,43 +1,6 @@
 import React, { useState, useEffect, useReducer } from "react";
+import { planets } from "../components/Constant";
 
-const planets = [
-  {
-    id: "01",
-    destination: "MOON",
-    description:
-      "See our planet as you’ve never seen it before. A perfect relaxing trip away to help regain perspective and come back refreshed. While you’re there, take in some history by visiting the Luna 2 and Apollo 11 landing sites.",
-    distance: "384,400 KM",
-    travelTime: "3 DAYS",
-    img: "/assets/destination/image-moon.webp",
-  },
-  {
-    id: "02",
-    destination: "MARS",
-    description:
-      "Don’t forget to pack your hiking boots. You’ll need them to tackle Olympus Mons, the tallest planetary mountain in our solar system. It’s two and a half times the size of Everest!",
-    distance: "225 MIL. KM",
-    travelTime: "9 MONTHS",
-    img: "/assets/destination/image-mars.webp",
-  },
-  {
-    id: "03",
-    destination: "EUROPA",
-    description:
-      "The smallest of the four Galilean moons orbiting Jupiter, Europa is a winter lover’s dream. With an icy surface, it’s perfect for a bit of ice skating, curling, hockey, or simple relaxation in your snug wintery cabin.",
-    distance: "628 MIL. KM",
-    travelTime: "3 YEARS",
-    img: "/assets/destination/image-europa.webp",
-  },
-  {
-    id: "04",
-    destination: "TITAN",
-    description:
-      "The only moon known to have a dense atmosphere other than Earth, Titan is a home away from home (just a few hundred degrees colder!). As a bonus, you get striking views of the Rings of Saturn.",
-    distance: "1.6 BIL. KM",
-    travelTime: "7 YEARS",
-    img: "/assets/destination/image-titan.webp",
-  },
-];
 export default function Destination() {
   const [selectedIndex, setSelectedIndex] = useState(0);
   const ref = [];
