+++ src/NewPAge/TechnologyPage.js
@@ -1,30 +1,5 @@
 import React, { useState, useEffect, useReducer } from "react";
-const technologies = [
-  {
-    id: "01",
-    name: "LAUNCH VEHICLE",
-    about:
-      "A launch vehicle or carrier rocket is a rocket-propelled vehicle used to carry a payload from Earth's surface to space, usually to Earth orbit or beyond. Our WEB-X carrier rocket is the most powerful in operation. Standing 150 metres tall, it's quite an awe-inspiring sight on the launch pad!",
-    landscape: "/assets/technology/image-launch-vehicle-landscape.jpg",
-    portrait: "/assets/technology/image-launch-vehicle-portrait.jpg",
-  },
-  {
-    id: "02",
-    name: "SPACEPORT",
-    about:
-      "A spaceport or cosmodrome is a site for launching (or receiving) spacecraft, by analogy to the seaport for ships or airport for aircraft. Based in the famous Cape Canaveral, our spaceport is ideally situated to take advantage of the Earth’s rotation for launch.",
-    landscape: "/assets/technology/image-spaceport-landscape.jpg",
-    portrait: "/assets/technology/image-spaceport-portrait.jpg",
-  },
-  {
-    id: "03",
-    name: "SPACE CAPSULE",
-    about:
-      "A space capsule is an often-crewed spacecraft that uses a blunt-body reentry capsule to reenter the Earth's atmosphere without wings. Our capsule is where you'll spend your time during the flight. It includes a space gym, cinema, and plenty of other activities to keep you entertained.",
-    landscape: "/assets/technology/image-space-capsule-landscape.jpg",
-    portrait: "/assets/technology/image-space-capsule-portrait.jpg",
-  },
-];
+import { technologies } from "../components/Constant";
 
 export default function TechnolohyPage() {
   const [selectedIndex, setSelectedIndex] = useState(0);
