+++ src/components/Constant.js
@@ -1,0 +1,129 @@
+const navList = [
+  {
+    id: "00",
+    to: "/",
+    title: "HOME",
+  },
+  {
+    id: "01",
+    to: "/destination",
+    title: "DESTINATION",
+  },
+  {
+    id: "02",
+    to: "/crew",
+    title: "CREW",
+  },
+  {
+    id: "03",
+    to: "/technology",
+    title: "TECHNOLOGY",
+  },
+  {
+    id: "04",
+    to: "/mission",
+    title: "MISSION",
+  },
+];
+
+const crews = [
+  {
+    id: "01",
+    position: "COMMANDER",
+    name: "DOUGLAS HURLEY",
+    about:
+      "Douglas Gerald Hurley is an American engineer, former Marine Corps pilot and former NASA astronaut. He launched into space for the third time as commander of Crew Dragon Demo-2.",
+    photo: "/assets/crew/image-douglas-hurley.png",
+  },
+  {
+    id: "02",
+    position: "MISSION SPECIALIST",
+    name: "MARK SHUTTLEWORTH",
+    about:
+      "Mark Richard Shuttleworth is the founder and CEO of Canonical, the company behind the Linux-based Ubuntu operating system. Shuttleworth became the first South African to travel to space as a space tourist.",
+    photo: "/assets/crew/image-mark-shuttleworth.webp",
+  },
+  {
+    id: "03",
+    position: "PILOT",
+    name: "VICTOR GLOVER",
+    about:
+      "Pilot on the first operational flight of the SpaceX Crew Dragon to the International Space Station. Glover is a commander in the U.S. Navy where he pilots an F/A-18.He was a crew member of Expedition 64, and served as a station systems flight engineer. ",
+    photo: "/assets/crew/image-victor-glover.webp",
+  },
+  {
+    id: "04",
+    position: "FlIGHT ENGINEER",
+    name: "ANOUSHEH ANSARI",
+    about:
+      "Anousheh Ansari is an Iranian American engineer and co-founder of Prodea Systems. Ansari was the fourth self-funded space tourist, the first self-funded woman to fly to the ISS, and the first Iranian in space. ",
+    photo: "/assets/crew/image-anousheh-ansari.webp",
+  },
+];
+
+const planets = [
+  {
+    id: "01",
+    destination: "MOON",
+    description:
+      "See our planet as you’ve never seen it before. A perfect relaxing trip away to help regain perspective and come back refreshed. While you’re there, take in some history by visiting the Luna 2 and Apollo 11 landing sites.",
+    distance: "384,400 KM",
+    travelTime: "3 DAYS",
+    img: "/assets/destination/image-moon.webp",
+  },
+  {
+    id: "02",
+    destination: "MARS",
+    description:
+      "Don’t forget to pack your hiking boots. You’ll need them to tackle Olympus Mons, the tallest planetary mountain in our solar system. It’s two and a half times the size of Everest!",
+    distance: "225 MIL. KM",
+    travelTime: "9 MONTHS",
+    img: "/assets/destination/image-mars.webp",
+  },
+  {
+    id: "03",
+    destination: "EUROPA",
+    description:
+      "The smallest of the four Galilean moons orbiting Jupiter, Europa is a winter lover’s dream. With an icy surface, it’s perfect for a bit of ice skating, curling, hockey, or simple relaxation in your snug wintery cabin.",
+    distance: "628 MIL. KM",
+    travelTime: "3 YEARS",
+    img: "/assets/destination/image-europa.webp",
+  },
+  {
+    id: "04",
+    destination: "TITAN",
+    description:
+      "The only moon known to have a dense atmosphere other than Earth, Titan is a home away from home (just a few hundred degrees colder!). As a bonus, you get striking views of the Rings of Saturn.",
+    distance: "1.6 BIL. KM",
+    travelTime: "7 YEARS",
+    img: "/assets/destination/image-titan.webp",
+  },
+];
+
+const technologies = [
+  {
+    id: "01",
+    name: "LAUNCH VEHICLE",
+    about:
+      "A launch vehicle or carrier rocket is a rocket-propelled vehicle used to carry a payload from Earth's surface to space, usually to Earth orbit or beyond. Our WEB-X carrier rocket is the most powerful in operation. Standing 150 metres tall, it's quite an awe-inspiring sight on the launch pad!",
+    landscape: "/assets/technology/image-launch-vehicle-landscape.jpg",
+    portrait: "/assets/technology/image-launch-vehicle-portrait.jpg",
+  },
+  {
+    id: "02",
+    name: "SPACEPORT",
+    about:
+      "A spaceport or cosmodrome is a site for launching (or receiving) spacecraft, by analogy to the seaport for ships or airport for aircraft. Based in the famous Cape Canaveral, our spaceport is ideally situated to take advantage of the Earth’s rotation for launch.",
+    landscape: "/assets/technology/image-spaceport-landscape.jpg",
+    portrait: "/assets/technology/image-spaceport-portrait.jpg",
+  },
+  {
+    id: "03",
+    name: "SPACE CAPSULE",
+    about:
+      "A space capsule is an often-crewed spacecraft that uses a blunt-body reentry capsule to reenter the Earth's atmosphere without wings. Our capsule is where you'll spend your time during the flight. It includes a space gym, cinema, and plenty of other activities to keep you entertained.",
+    landscape: "/assets/technology/image-space-capsule-landscape.jpg",
+    portrait: "/assets/technology/image-space-capsule-portrait.jpg",
+  },
+];
+export { navList, crews, planets, technologies };
